api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aerich==0.7.1
    # via -r requirements.in
aiosqlite==0.17.0
    # via tortoise-orm
anyio==3.6.2
    # via
    #   starlette
    #   watchfiles
asyncpg==0.27.0
    # via -r requirements.in
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   aerich
    #   uvicorn
dictdiffer==0.9.0
    # via aerich
fastapi==0.92.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
httptools==0.5.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
iso8601==1.1.0
    # via tortoise-orm
psycopg[binary,pool]==3.0.12
    # via tortoise-orm
psycopg-binary==3.0.12
    # via psycopg
psycopg-pool==3.1.6
    # via psycopg
pyaes==1.6.1
    # via pyrogram
pydantic==1.10.5
    # via
    #   aerich
    #   fastapi
pyjwt==2.6.0
    # via -r requirements.in
pypika-tortoise==0.1.6
    # via tortoise-orm
pyrogram==2.0.100
    # via -r requirements.in
pysocks==1.7.1
    # via pyrogram
python-dotenv==1.0.0
    # via uvicorn
pytz==2022.7.1
    # via tortoise-orm
pyyaml==6.0
    # via uvicorn
requests==2.28.2
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
starlette==0.25.0
    # via fastapi
tomlkit==0.11.6
    # via aerich
tortoise-orm[psycopg]==0.19.3
    # via
    #   -r requirements.in
    #   aerich
typing-extensions==4.5.0
    # via
    #   aiosqlite
    #   psycopg-pool
    #   pydantic
urllib3==1.26.14
    # via requests
uvicorn[standard]==0.20.0
    # via -r requirements.in
uvloop==0.17.0
    # via uvicorn
watchfiles==0.18.1
    # via uvicorn
websockets==10.4
    # via uvicorn
